refactor: move topic pane files into own directory
